fix: cleanup and tests passing
