+++ hyperdash/monitor.py
@@ -1,9 +1,7 @@
 # Python 2/3 compatibility
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-import logging
 import sys
-
 import uuid
 
 from .client import HDClient
@@ -11,21 +9,7 @@ from .code_runner import CodeRunner
 from .hyper_dash import HyperDash
 from .io_buffer import IOBuffer
 from .server_manager import ServerManagerHTTP
-
-def get_logger(model_name, current_sdk_run_uuid, stdout_buffer):
-   # Include the model_name/UUID in the logger name to make
-    # sure that its always distinct, even if multiple runs
-    # of the same model are happening at the same time in
-    # different threads
-    logger = logging.getLogger(
-        "{}-{}".format(model_name, current_sdk_run_uuid))
-    # Remove any existing log handlers so it doesn't double log
-    logger.handlers = []
-    # Don't propagate to the root logger
-    logger.propagate = False
-    logger.setLevel(logging.INFO)
-    logger.addHandler(logging.StreamHandler(stdout_buffer))
-    return logger
+from .utils import get_logger
 
 
 def monitor(model_name, api_key_getter=None, capture_io=True):
